RadialBasisFunctions

Documentation for RadialBasisFunctions.

RadialBasisFunctions.AbstractPHS

RadialBasisFunctions.AbstractRadialBasis

RadialBasisFunctions.Gaussian

RadialBasisFunctions.Laplacian

RadialBasisFunctions.MonomialBasis

RadialBasisFunctions.Partial

RadialBasisFunctions.RadialBasisOperator

RadialBasisFunctions.autoselect_k

RadialBasisFunctions.AbstractPHS — Type

abstract type AbstractPHS <: AbstractRadialBasis

Supertype of all Polyharmonic Splines.

RadialBasisFunctions.AbstractRadialBasis — Type

abstract type AbstractRadialBasis end

RadialBasisFunctions.Gaussian — Type

struct Gaussian{T,D<:Int} <: AbstractRadialBasis

Gaussian radial basis function:$ϕ(r) = e^{-(ε r)^2}$

RadialBasisFunctions.Laplacian — Type

Laplacian <: ScalarValuedOperator

Builds an operator for the sum of the second derivatives w.r.t. each independent variable.

RadialBasisFunctions.MonomialBasis — Type

struct MonomialBasis{T<:Int,B<:Function}

Multivariate Monomial basis. n ∈ N: length of array, i.e., x ∈ Rⁿ deg ∈ N: degree

RadialBasisFunctions.Partial — Type

Partial <: ScalarValuedOperator

Builds an operator for a first order partial derivative.

RadialBasisFunctions.RadialBasisOperator — Type

struct RadialBasisOperator <: AbstractRadialBasisOperator

Operator of data using a radial basis with potential monomial augmentation.

RadialBasisFunctions.autoselect_k — Method

autoselect_k(data::Vector, basis<:AbstractRadialBasis)

See Bayona, 2017 - https://doi.org/10.1016/j.jcp.2016.12.008