Exported Functions

RadialBasisFunctions.PHS — Method

function PHS(n::T=3; poly_deg::T=2) where {T<:Int}

Convienience contructor for polyharmonic splines.

RadialBasisFunctions.directional — Method

function directional(data, eval_points, v, basis; k=autoselect_k(data, basis))

Builds a RadialBasisOperator where the operator is the directional derivative, Directional.

RadialBasisFunctions.directional — Method

function directional(data, v, basis; k=autoselect_k(data, basis))

Builds a RadialBasisOperator where the operator is the directional derivative, Directional.

RadialBasisFunctions.gradient — Method

function gradient(data, basis; k=autoselect_k(data, basis))

Builds a RadialBasisOperator where the operator is the gradient, Gradient.

RadialBasisFunctions.gradient — Method

function gradient(data, eval_points, basis; k=autoselect_k(data, basis))

Builds a RadialBasisOperator where the operator is the gradient, Gradient. The resulting operator will only evaluate at eval_points.

RadialBasisFunctions.partial — Method

function partial(data, eval_points, order, dim, basis; k=autoselect_k(data, basis))

Builds a RadialBasisOperator where the operator is the partial derivative, Partial. The resulting operator will only evaluate at eval_points.

RadialBasisFunctions.partial — Method

function partial(data, order, dim, basis; k=autoselect_k(data, basis))

Builds a RadialBasisOperator where the operator is the partial derivative, Partial, of order with respect to dim.

RadialBasisFunctions.regrid — Method

function regrid(data, eval_points, basis=PHS(3; poly_deg=2); k=autoselect_k(data, basis))

Builds a RadialBasisOperator where the operator is an regrid from one set of points to another, data -> eval_points.

RadialBasisFunctions.∂virtual — Method

function ∂virtual(data, eval_points, dim, Δ, basis; k=autoselect_k(data, basis))

Builds a virtual RadialBasisOperator whichi will be evaluated at eval_points where the operator is the partial derivative with respect to dim. Virtual operators interpolate the data to structured points at a distance Δ for which standard finite difference formulas can be applied.

RadialBasisFunctions.∂virtual — Method

function ∂virtual(data, dim, Δ, basis; k=autoselect_k(data, basis))

Builds a virtual RadialBasisOperator whichi will be evaluated at the input points (data) where the operator is the partial derivative with respect to dim. Virtual operators interpolate the data to structured points at a distance Δ for which standard finite difference formulas can be applied.

RadialBasisFunctions.AbstractPHS — Type

abstract type AbstractPHS <: AbstractRadialBasis

Supertype of all Polyharmonic Splines.

RadialBasisFunctions.AbstractRadialBasis — Type

abstract type AbstractRadialBasis <: AbstractBasis end

RadialBasisFunctions.Directional — Type

Directional <: ScalarValuedOperator

Operator for the directional derivative, or the inner product of the gradient and a direction vector.

RadialBasisFunctions.Gaussian — Type

struct Gaussian{T,D<:Int} <: AbstractRadialBasis

Gaussian radial basis function:$ϕ(r) = e^{-(ε r)^2}$

RadialBasisFunctions.Gradient — Type

Gradient <: VectorValuedOperator

Builds an operator for the gradient of a function.

RadialBasisFunctions.Interpolator — Type

struct Interpolator

Construct a radial basis interpolation.

RadialBasisFunctions.Interpolator — Method

function Interpolator(x, y, basis::B=PHS())

Construct a radial basis interpolator.

RadialBasisFunctions.Laplacian — Type

Laplacian <: ScalarValuedOperator

Operator for the sum of the second derivatives w.r.t. each independent variable.

RadialBasisFunctions.MonomialBasis — Type

struct MonomialBasis{Dim,Deg} <: AbstractBasis

Dim dimensional monomial basis of order Deg.

RadialBasisFunctions.PHS1 — Type

struct PHS1{T<:Int} <: AbstractPHS

Polyharmonic spline radial basis function:$ϕ(r) = r$

RadialBasisFunctions.PHS3 — Type

struct PHS3{T<:Int} <: AbstractPHS

Polyharmonic spline radial basis function:$ϕ(r) = r^3$

RadialBasisFunctions.PHS5 — Type

struct PHS5{T<:Int} <: AbstractPHS

Polyharmonic spline radial basis function:$ϕ(r) = r^5$

RadialBasisFunctions.PHS7 — Type

struct PHS7{T<:Int} <: AbstractPHS

Polyharmonic spline radial basis function:$ϕ(r) = r^7$

RadialBasisFunctions.Partial — Type

Partial <: ScalarValuedOperator

Builds an operator for a first order partial derivative.

RadialBasisFunctions.RadialBasisOperator — Type

struct RadialBasisOperator

Operator of data using a radial basis with potential monomial augmentation.

RadialBasisFunctions.Regrid — Type

Regrid

Builds an operator for interpolating from one set of points to another.

Private

RadialBasisFunctions.autoselect_k — Method

autoselect_k(data::Vector, basis<:AbstractRadialBasis)

See Bayona, 2017 - https://doi.org/10.1016/j.jcp.2016.12.008

RadialBasisFunctions.AbstractBasis — Type

abstract type AbstractBasis end