Exported Functions

RadialBasisFunctions.PHS — Method

function PHS(n::T=3; poly_deg::T=2) where {T<:Int}

Convienience contructor for polyharmonic splines.

RadialBasisFunctions.AbstractPHS — Type

abstract type AbstractPHS <: AbstractRadialBasis

Supertype of all Polyharmonic Splines.

RadialBasisFunctions.AbstractRadialBasis — Type

abstract type AbstractRadialBasis end

RadialBasisFunctions.Gaussian — Type

struct Gaussian{T,D<:Int} <: AbstractRadialBasis

Gaussian radial basis function:$ϕ(r) = e^{-(ε r)^2}$

RadialBasisFunctions.Gradient — Type

Gradient <: VectorValuedOperator

Builds an operator for the gradient of a function.

RadialBasisFunctions.Laplacian — Type

Laplacian <: ScalarValuedOperator

Builds an operator for the sum of the second derivatives w.r.t. each independent variable.

RadialBasisFunctions.MonomialBasis — Type

struct MonomialBasis{T<:Int,B<:Function}

Multivariate Monomial basis. n ∈ N: length of array, i.e., x ∈ Rⁿ deg ∈ N: degree

RadialBasisFunctions.PHS1 — Type

struct PHS1{T<:Int} <: AbstractPHS

Polyharmonic spline radial basis function:$ϕ(r) = r$

RadialBasisFunctions.PHS3 — Type

struct PHS3{T<:Int} <: AbstractPHS

Polyharmonic spline radial basis function:$ϕ(r) = r^3$

RadialBasisFunctions.PHS5 — Type

struct PHS5{T<:Int} <: AbstractPHS

Polyharmonic spline radial basis function:$ϕ(r) = r^5$

RadialBasisFunctions.PHS7 — Type

struct PHS7{T<:Int} <: AbstractPHS

Polyharmonic spline radial basis function:$ϕ(r) = r^7$

RadialBasisFunctions.Partial — Type

Partial <: ScalarValuedOperator

Builds an operator for a first order partial derivative.

RadialBasisFunctions.RadialBasisInterp — Type

struct RadialBasisInterp

Type for radial basis interpolation.

RadialBasisFunctions.RadialBasisInterp — Method

function RadialBasisInterp(x, y, basis::B=PHS())

Construct a radial basis interpolator.

RadialBasisFunctions.RadialBasisOperator — Type

struct RadialBasisOperator

Operator of data using a radial basis with potential monomial augmentation.

Private

RadialBasisFunctions.autoselect_k — Method

autoselect_k(data::Vector, basis<:AbstractRadialBasis)

See Bayona, 2017 - https://doi.org/10.1016/j.jcp.2016.12.008